refactor (Organiza archivos): Organiza archivos y carpetas. Crea carpeta /services. Mueve archivos a /services y /utils. Crea barrels en carpetas señaladas
